pyproject.toml
[project]
name = "image-gremlin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "pillow>=12.1.0",
]
